requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.27
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
Markupsafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
crewai==0.126.0